add: logica para productos, pedidos y vendedores

+++ Infraestructure/Models/MetodoPago.cs
@@ -23,7 +23,7 @@ namespace Infraestructure.Models
         public int id { get; set; }
         public int usuarioId { get; set; }
         public string nombre { get; set; }
-        public int numero { get; set; }
+        public long numero { get; set; }
     
         public virtual Usuario Usuario { get; set; }
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
